Refactor : 로그인 확인 여부, 유저 정보 확인하는 로직을 커스텀 훅으로 분리
